requirements.txt
streamlit
firebase-admin
pandas
python-dotenv
google-generativeai